Split different utils pipe into ipc module
The pipe module was becoming too large and with containg different
utilities and structs that are conceptually different. This patch just
reorganizes the content of pipe into an ipc module.
