+++ northstar/src/runtime/mod.rs
@@ -31,9 +31,9 @@ pub mod config;
 mod console;
 mod debug;
 mod error;
+mod ipc;
 mod key;
 mod mount;
-mod pipe;
 pub(crate) mod process;
 mod repository;
 mod state;
